+++ sr_robot_launch/CMakeLists.txt
@@ -179,7 +179,14 @@ if (CATKIN_ENABLE_TESTING)
 #    add_rostest(test/test_hand_joint_movement.test ARGS hand_type:=hand_lite hand_id:=lh)
 #    add_rostest(test/test_hand_joint_movement.test ARGS hand_type:=hand_extra_lite hand_id:=rh)
 #    add_rostest(test/test_hand_joint_movement.test ARGS hand_type:=hand_extra_lite hand_id:=lh)
-    add_rostest(test/test_hand_and_arm_sim.test)
+    add_rostest(test/test_ur10_right.test)
+    add_rostest(test/test_ur10_right.test ARGS scene:=true)
+    add_rostest(test/test_ur10_left.test)
+    add_rostest(test/test_ur10_left.test ARGS scene:=true)
+    add_rostest(test/test_ur5_right.test)
+    add_rostest(test/test_ur5_right.test ARGS scene:=true)
+    add_rostest(test/test_ur5_left.test)
+    add_rostest(test/test_ur5_left.test ARGS scene:=true)
 
 endif ()
