+++ sr_robot_launch/test/test_launch_file.launch
@@ -1,16 +1,18 @@
 <launch>
-     <!-- Find launch file-->
-     <arg name="launch_file" default="sr_right_ur10arm_hand.launch"/>
-     <param name="/test_sim/launch_file" value="$( arg launch_file)"/>
+     <!-- Launch with or without scene-->
      <arg name="scene" default="false"/>
-     <param name="/test_sim/scene" value="$( arg scene)"/>
-     <arg name="start_home" default="true"/>
-     <param name="/test_sim/start_home" value="$( arg start_home)"/>
+     <param name="/test_sim/scene" value="$( arg scene)"/> 
 
-<include file="$(find sr_robot_launch)/launch/$(arg launch_file)">
+     <!-- Choose hand type -->
+     <arg name="hand_type" default="hand_e"/>
+     <param name="/test_sim/hand_type" value="$( arg hand_type)"/>
+
+<include file="$(find sr_robot_launch)/launch/sr_right_ur10arm_hand.launch">
      <arg name="sim" value="true"/>
      <arg name="gui" value="true"/>
      <arg name="scene" value="$( arg scene)"/>
-     <arg name="start_home" value="$( arg start_home)"/>
+     <!-- <arg name="side" value="right"/> -->
+     <arg name="hand_type" value="$( arg hand_type)"/>
+     <!-- <arg name="robot_model" value="ur10"/> -->
 </include>
 </launch>
