--- sr_robot_launch/test/test_launch_file_2.launch
@@ -1,18 +1,0 @@
-<launch>
-     <!-- Launch with or without scene-->
-     <arg name="scene" default="false"/>
-     <param name="/test_sim/scene" value="$( arg scene)"/> 
-
-     <!-- Choose hand type -->
-     <arg name="hand_type" default="hand_e"/>
-     <param name="/test_sim/hand_type" value="$( arg hand_type)"/>
-
-<include file="$(find sr_robot_launch)/launch/sr_right_ur10arm_hand.launch">
-     <arg name="sim" value="true"/>
-     <arg name="gui" value="true"/>
-     <arg name="scene" value="$( arg scene)"/>
-     <!-- <arg name="side" value="right"/> -->
-     <arg name="hand_type" value="$( arg hand_type)"/>
-     <!-- <arg name="robot_model" value="ur10"/> -->
-</include>
-</launch>
