+++ sr_robot_launch/test/test_ur10_left.test
@@ -3,7 +3,7 @@
      <arg name="scene" default="false"/>
      <arg name="hand_type" value="hand_e"/>
 
-<include file="$(find sr_robot_launch)/launch/sr_right_ur10arm_hand.launch">
+<include file="$(find sr_robot_launch)/launch/sr_left_ur10arm_hand.launch">
      <arg name="sim" value="true"/>
      <arg name="gui" value="false"/>
      <arg name="scene" value="$( arg scene)"/>
