+++ sr_robot_launch/test/test_ur5_left.test
@@ -1,9 +1,9 @@
 <launch>
      <!-- Find launch file-->
      <arg name="scene" default="false"/>
-     <arg name="hand_type" value="hand_e"/>
+     <arg name="hand_type" value="hand_lite"/>
 
-<include file="$(find sr_robot_launch)/launch/sr_right_ur10arm_hand.launch">
+<include file="$(find sr_robot_launch)/launch/sr_left_ur5arm_hand.launch">
      <arg name="sim" value="true"/>
      <arg name="gui" value="false"/>
      <arg name="scene" value="$( arg scene)"/>
